antigravity/requirements.txt
feedparser==6.0.11
requests==2.31.0
Pillow==10.4.0
google-api-python-client==2.142.0
google-auth==2.34.0
python-dateutil==2.9.0
google-generativeai==0.8.3
beautifulsoup4==4.12.3
